Adjust location of tests
